+++ Data/SBV/Provers/Boolector.hs
@@ -36,6 +36,7 @@ boolector = SMTSolver {
                               , supportsDoubles            = False
                               , supportsOptimization       = False
                               , supportsPseudoBooleans     = False
+                              , supportsUnsatCores         = True
                               }
          }
  where addTimeOut o i | i < 0 = error $ "Boolector: Timeout value must be non-negative, received: " ++ show i